Traktor: Fixed spline extrude without baking. A lot of other tweaks.

+++ code/Terrain/Editor/ElevateBrush.h
@@ -48,7 +48,7 @@ private:
 	Ref< drawing::Image > m_splatImage;
 	int32_t m_radius;
 	const IFallOff* m_fallOff;
-	float m_strength;
+	double m_strength;
 	int32_t m_attribute;
 };
